content views - migrating to using CV version
Now each content view had version objects attached to it
that keep track of environments and repos of a particular
content view instance

Also, changed content view definition repo to be a many to many
association, as thats what it should be :)

+++ src/app/models/content_view.rb
@@ -17,20 +17,18 @@ class ContentView < ActiveRecord::Base
 
   belongs_to :content_view_definition
   belongs_to :organization
-  has_one :environment_default, :class_name => "KTEnvironment",
-    :inverse_of => :default_content_view,
-    :foreign_key => :default_content_view_id
 
-  has_many :environment_content_views
-  has_many :environments, :through => :environment_content_views,
-    :class_name => "KTEnvironment", :inverse_of=>:content_views
+  has_many :content_view_versions
+  alias :versions :content_view_versions
+
+  has_one :environment_default, :class_name => "KTEnvironment",
+      :inverse_of => :default_content_view,
+      :foreign_key => :default_content_view_id
 
 
   has_many :component_content_views
   has_many :composite_content_view_definitions,
     :through => :component_content_views, :source => "content_view_definition"
-
-  has_many :repositories, :dependent => :destroy
 
   has_many :changeset_content_views
   has_many :changesets, :through => :changeset_content_views
@@ -40,11 +38,8 @@ class ContentView < ActiveRecord::Base
   validates :name, :presence => true, :katello_name_format => true
   validates :organization_id, :presence => true
 
-  scope :default, joins('left outer join environments on content_views.id = environments.default_content_view_id').
-      where('environments.default_content_view_id is not null')
-  scope :non_default, joins('left outer join environments on content_views.id = environments.default_content_view_id').
-        where('environments.default_content_view_id is null')
-
+  scope :default, where(:default=>true)
+  scope :non_default, where(:default=>false)
 
   def as_json(options = {})
     result = self.attributes
@@ -57,9 +52,21 @@ class ContentView < ActiveRecord::Base
     result
   end
 
-  #is this content view a default
-  def default?
-    !self.environment_default.nil?
+  def environments
+    KTEnvironment.joins(:content_view_versions).where('content_view_versions.content_view_id' => self.id)
+  end
+
+  def version(env)
+    self.versions.in_environment(env).first
+  end
+
+  def repos(env)
+    version = version(env)
+    if version
+      version.repositories.in_environment(env)
+    else
+      []
+    end
   end
 
   def promote(from_env, to_env)
@@ -67,25 +74,22 @@ class ContentView < ActiveRecord::Base
     #remove this when refresh is supported
     raise "Cannot promote to #{to_env.name}, view already exist there and refreshing not supported." if self.environments.include?(to_env)
 
-    self.environments << to_env
-    self.save!
+    version = self.version(from_env)
+    version.environments << to_env
+    version.save!
     tasks = []
-    self.repositories.each do |repo|
-      clone = repo.create_clone(to_env, self, to_env.default_content_view)
+    self.repos(from_env).each do |repo|
+      clone = repo.create_clone(to_env, self)
       tasks << repo.clone_contents(clone)
     end
     tasks
   end
 
   def delete(from_env)
-    raise "Cannot delete from #{from_env.name}, view does not exist there." if !self.environments.include?(from_env)
-    self.environments.delete(from_env)
-    self.repositories.in_environment(from_env).each{|r| r.destroy}
-    if self.environments.empty?
-      self.destroy
-    else
-      self.save!
-    end
+    version = self.version(from_env)
+    raise "Cannot delete from #{from_env.name}, view does not exist there." if version.nil?
+    version.delete(from_env)
+    self.destroy if self.versions.empty?
   end
 
 end
